linux-osf/qcom.cfg: update settings for DB820C and DB410C
During the latest kernel rebase, we avoided the patches from the
QCOM landing team which added configs to the arm64 defconfig.
Instead, add the config changes in our distro fragment for easier
maintenance.  This applies to GLINK RPMSG and UFS enablement
configs.

CONFIG_QCOM_APCS_IPC is a new mailbox driver implemented upstream
and is now a requirement for proper operation of DB820C.

QCOM_EBI2 is for 32-bit QCOM devices (not DB820C or DB410C).

Signed-off-by: Michael Scott <michael@opensourcefoundries.com>
Signed-off-by: Ricardo Salveti <ricardo@opensourcefoundries.com>