docker: use linkmode external when building docker-proxy
Use the external linkmode when building docker-proxy so it gets the right
linker path when the target and host architectures are the same. This
is also done by the internal install-binaries.sh script, normally used
when building docker from scratch.

Issue found when building docker for intel-corei7-64 on an Intel box,
which had docker-proxy linker path pointing to
/lib64/ld-linux-x86-64.so.2 instead of /lib/ld-linux-x86-64.so.2.

Signed-off-by: Ricardo Salveti <ricardo@opensourcefoundries.com>