initramfs-framework: add specific ostree script for freedom-u540 and qemuriscv64
There is no generic bootloader support for both targets so define a
ostree specific initrd script that tries to find out the default ostree
hash to be used, as that is not given by default to the kernel.

To be reverted once the bootloader is able fetch the ostree hash to use
from the storage device.

Signed-off-by: Ricardo Salveti <ricardo@opensourcefoundries.com>