lmp-gateway-image: add haveged to remedy low-entropy
On headless embedded systems entropy is often rather low. This
leads to a delayed initialization of the random pool ("crng init
done"), e.g. on a Colibri iMX7:
  # dmesg | grep random
  [    0.000000] random: fast init done
  [    6.944298] systemd[1]: Initializing machine ID from random generator.
  [  143.523689] random: crng init done

This also delays docker startup since docker seems to make use of
the getrandom() syscall, which stalls until crng init is done.

Add haveged which uses the HAVEGE algorithm and makes sure that
enough entropy is available early on.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>