lmp-image-gateway: set IMAGE_OVERHEAD_FACTOR to 2.3 for full image update
By default IMAGE_OVERHEAD_FACTOR is set to 1.3 which is good enough for
a fully working system, but not necessarily enough for an image update.

This is not an issue for systems using a storage device that is already
at least twice the image size, as resizefs gets called during first
boot, but it is better to force a minimal image size for a better user
experience (e.g. qemu/vbox users).

Only changing lmp-gateway-image for now as we need a kernel modules/size
refactoring before applying a similar logic in the mini-image, otherwise
users are not going to be able to run the image with a storage device
with less than 1GB.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>