docker: improve system dockerd config logic
While changing the default systemd job to give dockerd a default system
daemon.json configuration file works as expected, it ends up being
harder for the user to replace with a custom config.

Update dockerd to only use the system provided config
(/usr/lib/docker/daemon.json) when the default, or the one provided via
the command line, cannot be found in the system.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>