linux-firmware: update WL8 service pack to v4.3
Updated Fixes in Service Pack 4.3

- Public Key Validation for BR/EDR Secure Simple Pairing (SSP)
This service pack validates the public key received over the air
when pairing a new BR/EDR device using the Secure Simple Pairing (SSP).
This security enhancement provides additional protection against certain
man-in-the-middle type of attacks.

- Fix Disconnection due to peer sniff attempt of 1
The BT classic connection may get disconnected with error code
(0x08 - Connection Timeout) when the peer device sends sniff request
with sniff attempt 1 and the local WL18xx controller is the slave in
the connection. This service pack fixes this bug by renegotiating sniff
attempt of 4 with the peer device so the connection timeout does not happen.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>