lmp-machine-custom: rpi: use ttyS0 by default when uart is enabled
serial0 only works correctly when the proprietary boot firmware is the
one responsible for loading and parsing the kernel command line, as
it gets replaced with ttyS0 or ttyAMA0 based on the target.

Since we are using u-boot as the main bootloader, use the correct console
device directly for the targets we support (rpi 0-wifi and 3/3+).

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>