procps: add updated sysctl.conf to fix rp_filter issue
OE has a very outdated sysctl.conf file in the procps package:
https://git.openembedded.org/openembedded-core/commit/?id=8a9b9a323f4363e27138077e3e3dce8139a36708
(circa 2014)

The origins of this file is quite unknown and it's causing a routing
issue when both wifi and ethernet are enabled due to the following:
- upstream change in NetworkManager during 1.16 cycle removes the
  dynamic rp_filter setting change when more than one interface
  is enabled.
- legacy setting in sysctl.conf sets rp_filter to 1 which blocks
  packets with different inbound and outbound addresses.

Documentation of rp_filter setting from kernel.org:

rp_filter - INTEGER
0 - No source validation.
1 - Strict mode as defined in RFC3704 Strict Reverse Path
    Each incoming packet is tested against the FIB and if the interface
    is not the best reverse path the packet check will fail.
    By default failed packets are discarded.
2 - Loose mode as defined in RFC3704 Loose Reverse Path
    Each incoming packet's source address is also tested against the FIB
    and if the source address is not reachable via any interface
    the packet check will fail.

This patch updates the sysctl.conf file to procps v3.3.15 which doesn't
set the rp_filter mode explicity.

NOTE: The kernel/pid_max=10000 setting has been commented out as this
may not be desired by default.

Signed-off-by: Michael Scott <mike@foundries.io>