curl: also set default ca-path
Without setting a default ca-path curl fails to find the certificates
available on the system when another certificate gets set and used by
the user (as it shares the same option).

Having a default ca-path allows curl to use both the certificate
provided by the user and the certificates available on the system. This
option works with both OpenSSL (used by default with sota) and GnuTLS.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>