lmp: control SOTA BSP configs options
meta-updater defines a set of sota_${MACHINE}.bbclass files that are
used to change BSP related options in order to make the machine
compatible with sota, but those options are not necessarily the ones
desired by LMP, forcing us to define a series of sota_remove overrides.

Instead of letting meta-updater define the BSP options to be used by the
distro, force the default override to be lmp so we can define and control
the options that are relevant to lmp itself.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>