bsp: optee-os: imx8mmevk: disable DYN_SHM
DRAM is not properly initialized in OPTEE-OS on iMX targets yet, so
instead of allowing any dyn memory just force it to use the shared
memory area instead.

This is needed only for imx8mm as the device dtb (used by SPL) is not
available for OPTEE during initialization, so it can't simply just map
the memory node from the device tree when setting up the dyn area.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>