bsp: lmp-machine-custom: rpi: unset kbuild defconfig for cm3
Board config sets a default kbuild defconfig value, but this has a bad
side effect on every linux recipe as fragments are completely ignored by
the build process. Drop kbuild defconfig settings for cm3 so it is able
to fetch and use the configs from the kernel meta repository instead.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>