docker-compose: Add support for compose apps
This adds a small out-of-tree patch to docker-compose that allows us
to "download" and Compose App. Some context about apps can be found
here:

 https://foundries.io/insights/2020/04/17/container-orchestration-v2/

We decided that "Bundles" was a bit confusing in the product since we
were using "Apps" everywhere. So its now called "compose apps" instead
of "compose bundles".

Signed-off-by: Andy Doan <andy@foundries.io>