bsp: imx8mmevk: add dts compatible with previous NXP BSP
Add base device tree what was previously used in the older NXP BSP for
imx8mmevk. This is required as previous images (using older u-boot) are
still going to search a fit config entry that is not available anymore
on the NXP 5.4-based BSP (as all the dts files were renamed).

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>