base: lmp-image-common: only increase overhead factor for QEMU targets
As a part of the LmP common image changes, we run a resize-helper which
expands the rootfs to fill the available space.  This usually takes care
of needing enough space on the rootfs to download a complete OTA image.
(about double the size at a minimum)

QEMU target images are not expanded past their original size during this
process and so we increased the overhead factor for all targets to 2.3.

This has 2 downsides:
1. Subscribers have a hard time overriding this setting
2. non-QEMU target wic images are needlessly larger than they should be
   due to the resize-helper.

By selectively setting an overridable value of 2.3 for QEMU targets,
we solve both of these issues.

Signed-off-by: Michael Scott <mike@foundries.io>