base: systemd: extend default fallback ntp servers
Extend the default list of fallback ntp server to include the ntp server
from cloudflare.com. This is done in a way to avoid relying on one
single domain, which might be blocked in certain networks.

This change also allows the user to easily overwrite the default
fallback via the DEF_FALLBACK_NTP_SERVERS variable.

Change based on a similar change done in meta-toradex-torizon.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>