base: docker-ce: restart in case of service failure
Docker is a crucial service for LMP, so make sure the service triggers a
reboot in case it fails to start, as then the system can automatically
perform a rollback in case of a bad update.

Also enable TasksMax=infinity since this is recommended by the Docker
team when using newer systemd version than 226.

Change based on a similar change done in meta-toradex-torizon.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>