base: docker-ce: depend on boot-complete.target
Make use of the boot-complete.target synchronization point. Order the
unit docker.service before this target to make its start necessary for
a boot to be considered successful.

This makes sure that aktualizr-lite.service (which is ordered after
boot-complete.target) only gets started if the docker.service is
successful, and hence a boot is only marked successful if Docker starts
and is functional.

Change based on a similar change done in meta-toradex-torizon.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>