base: linux-lmp-dev: add workaround for oe-core e684532
OE-Core e684532 changed the sccs parsing order, making defconfig be the
first one to be parsed, but the required elements check wasn't updated
to include the new sccs_defconfig variable, causing scc to be skipped
and resulting in a kernel build failure.

Workaround the issue by defining KERNEL_FEATURES with a common and basic
config fragment, as then elements will not be empty, causing scc to be
called as usual.

Can be reverted once the proper fix lands upstream.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>