base: lmp: extend bbmask for meta-raspberrypi
fw_env.config shouldn't be provided by libubootenv and we already have
the required changes in place for u-boot-fio, so mask the bbappend.

This same change was already reverted from meta-raspberrypi/master, just
not yet in dunfell.

This patch can be reverted once meta-raspberrypi/dunfell gets updated
with the same fix as done in master.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>