base: kernel-lmp-fitimage: add support for FIT_LOADABLES
Extend this class to handle FIT_LOADABLES as a list of deployed files
to be added to the kernel fitimage.  This allows subscribers to
retreive these files in u-boot where the contents are verified prior
to using.

Example would be to add the following line to machine.conf:
FIT_LOADABLES = "core0_m4_image.bin core1_m4_image.bin"

- A recipe would place those files in ${DEPLOYDIR} in do_deploy
- Then an additional do_compile[depends] is added to the kernel
  recipe such as:
  do_compile[depends] += "<recipe>:do_deploy"

Signed-off-by: Michael Scott <mike@foundries.io>