base: wireguard-module: set rprovides if kernel != linux-lmp
Now that linux-lmp has moved to 5.8, we can't have wireguard-module
being built as part of this recipe, so make sure there is no rprovides
in case the kernel used is linux-lmp.

This won't fix all the use cases we have, but it is the easiest way to
know when it should be built from the kernel or via this recipe (there
is no easy way to parse LINUX_VERSION during dependency checks).

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>