bsp: rpi: switch to firmware provided dtb/overlays
Prefer using the dtb that is provided by the proprietary rpi boot
firmware instead of managing and loading via kernel fit image as this is
the supported path by rpi upstream.

Firmware is now responsible for extending and adding nodes based on the
configuration defined in config.txt and the board rev/type, enough that
the default dtb provided by the kernel doesn't enable some basic
functionalities such as bluetooth and wifi anymore.

By switching to the supported way as defined by rpi upstream we can
reuse the workflow that is already documented and expected by rpi users,
reducing the maintenance work from our side. The downside is that the
user might have to manually update the boot firmware from time to time,
otherwise the runtime device tree might become incompatible with the
kernel used by the build.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>