bsp: trusted-firmware-a: corstone700: set bootargs/initrd in dtb
Set the expected bootargs and initrd address location in the corstone
fpga target as we require in LMP (rootfs located in a usb disk).

This won't be needed once u-boot is fully able to load the kernel/initrd
from the usb disk, and set up the right addresses and offsets.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>