base: linux-yocto: add generic appends from meta-virtualization
We mask the generic linux-%.bbappend file from meta-virtualization as
that ends up affecting every kernel flavour, which we don't really want.

The bad side effect is that linux-yocto users won't have support for the
required meta-virtualization fragment, so add a local append that is
just for linux-yocto recipes, based on the generic append from
meta-virtualization.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>