base: aktualizr: drop aktualizr-lite.path
While it is useful to have a path monitor in order to quickly start
aktualizr-lite after performing a registration (e.g.
lmp-device-register), it causes way too much confusion when the user is
trying to stop or manage the service, as this can bring unexpected
behavior.

A similar logic will be moved to lmp-device-register, to allow having a
similar behavior but just after registration.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>