bsp: lmp-machine-custom: intel-corei7-64: add support for wic-based installer
wic-based installer does not have the 4gb size limit from the standard
live/hddimg format, as the partition used to store the rootfs content is
based on ext4.

The only issue with the wic-based installer is that wic itself can only
consume one wks file, even if multiple are defined via WKS_FILES, which
forces the user to build the traditional image or the installer image,
not both.

To build the wic-based installer just set the WKS_FILES_<machine>_sota
varilable to image-efi-installer.wks.in (which can be set by the user
in local.conf or conf/machine/include/lmp-factory-custom.inc via their
own factory).

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>