base: mfgtool-files: name tarball with the machine name
Use the machine name as part of the base folder name for mfgtool-files
to make it more explict and also allow the user to easily manage
multiple tarballs without having to rename them.

In order to keep it compatible with the previous naming (from the
publishing perspective), a link is created without the machine name.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>