bsp: u-boot-ostree-scr-fit: fix altbootcmd for supported targets
Rollback logic via altbootcmd was only correct when called by the boot
script, as it had all the variables and environment required for it to
boot the previous image.

Now when u-boot calls altbootcmd automatically (after bootcount >
bootlimit), the boot script would not be called, which resulted in a
partial environment (lacking the actual uEnv.txt content from ostree).

Rollback still worked but it required an additional reboot cycle as the
initial trial would fail, but the following one would be correct as
rollbackenv would set rollback=1 and upgrade_available=0.

Fix the logic when called by u-boot by saving the resetvars and otenv
load process as part of the environment.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>