base: u-boot-ostree-scr/fit: add support for fit node separator
Add support to dynamically select the fit node separator used when
loading nodes from the kernel fitimage. This is sync with
kernel-lmp-fitimage.bbclass, and the same variable can be used to
control the behavior for both (which can be set per target/factory).

Boot.cmd.in is not created for the sed step at this point because
factory layers are all providing custom boot.cmd and it would require
manual renaming to avoid build failures. Since we will later switch to
remove unit address references entirely, we should be able to drop the
required sed in a later release.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>