bsp: u-boot-ostree-scr/fit: allow built time selection for fit node separator
As an effort to move away from using unit address when loading the FIT
components (u-boot CVE-2021-27138), allow the fit node separator to be
set during build time via the FIT_NODE_SEPARATOR variable.

This is required as the fit node separator from boot.cmd needs to be
fully in sync with what is set in the kernel fit, otherwise the
machine won't be able to boot the kernel fit.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>