base: linux-firmware: bring recipe from oe-core master cd273c611b
Temporarily bring the linux-firmware recipe from oe-core master
cd273c611b as a way to fix the packaging regression introduced when the
base recipe was updated to 20210208, which caused some packages to
depend on the entire linux-firmware meta package, causing a huge
increase in image size (over 600mb).

Can be reverted once a similar change lands in gatesgarth.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>