bsp: recipes-bsp: u-boot-fio-mfgtool: imx6ullevk: update config
The current mfgtool SPL for imx6ull is a bit unconventional due to
the limited OCRAM size.  It is loaded to RAM along with u-boot.itb
and then SDP jumps to SPL.

For ease of signing this SPL and standardization, let's re-adjust
it to work the same way as other SPL that are supported in LmP.
(SDP loads to RAM and then jumps, then SPL does it's own load of
u-boot.itb).

Also included in the cleanup:
- Proper size check to include room for i.MX signing data
- Drop SPL_RAM_SUPPORT
- Add SPL_USB_SDP_SUPPORT
- Drop SPL_DM for size reduction
- Loglevel 2 for size reduction
- Simplify config by removing redundant settings
- Drop RPMB to avoid mis-use on open HW

Signed-off-by: Michael Scott <mike@foundries.io>