bsp: opensbi: support fw_dynamic u-boot fit-image for freedom-u540
MACHINE conf for freedom-u540 specifies a default value for
RISCV_SBI_PAYLOAD.  The check for this in opensbi 0.8 recipe only
handles a None value where the FW_PAYLOAD_PATH is not configured
(not an empty "" value).

Fix that by removing @riscv_get_extra_oemake_image(d) from EXTRA_OEMAKE.

Also add a dependency to u-boot-fio for opensbi so that it can be
included in the u-boot fit-image.

Signed-off-by: Michael Scott <mike@foundries.io>