bsp: add machine configuration for imx8mmevk
The old imx8mmevk machine configuration was split and renamed to
imx8mm-lpddr4-evk and imx8mm-ddr4-evk in meta-freescale hardknott, so
add back the imx8mmevk machine configuration assuming imx8mm-lpddr4-evk
as the default option, which is the machine we support in lmp anyway.

This is done in a way to allow our users to transition to the new
machine names without breaking support.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>