bsp: ultrazed-lmp: do not override UBOOT_SUFFIX
Config sets u-boot to compile u-boot.bin instead of u-boot.itb, and if
we set UBOOT_SUFFIX to itb it will force OE to try to install the
incorrect file and fail during install/deploy.

Remove the UBOOT_SUFFIX override and let it with the original value
"bin" which is set for all zynqmp devices.

Later in the deploy step the final u-boot.itb file will be generated via
the uboot-fitimage.bbclass.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>