base: lmp-image-common: create fio (LMP_USER) group before the user
For useradd-staticids to work correctly, the group needs to be created
via the groupadd call (in order for it to find and match the group id),
so make sure to add the FIO_USER group before the user gets created.

This fixes a rootfs build failure when useradd-staticids is enabled.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>
(cherry picked from commit b1cc7e0d271a1f6f09b1772e277422a51b9e3382)