base: Remove prebuilt image logic
We are changing CI to trigger an "assemble-system-image" run that works
identical to our container builds. This gives us one code path that is
capable of dealing with our advanced tagging scenarios customers are
starting to hit.

Dunfell: at that point, preload app wasn't part of the backport. I just
add this commit because I saw that the previous commit already added
this if and this commit is removing it. So I decided to adapt and remove
the if.

Signed-off-by: Andy Doan <andy@foundries.io>
(cherry picked from commit 11187730a62cf1eff152746657ec29b73077cdb2)
Signed-off-by: Raul Muñoz <raul@foundries.io>