base: ostree: patch to setup a curl request timeout
Set a timeout for an overall request processing in order to avoid an
ostree pull/fetch API call or CLI command hang.
The timeout value can be specified via an env variable OSTREE_CURL_TIMEOUT,
if not defined, or equal 0 or it cannot be converted to a long type then
the default value is set.

Signed-off-by: Mike Sul <mike.sul@foundries.io>