bsp: u-boot-ostree-scr-fit: imx7ulpea-ucom: update loadaddr
With IMX_HAB we need to align loadaddr with the address used by fit,
otherwise bootm will fail as it looks for loadaddr by default when
verifying the image.

This is only an issue with the NXP BSP u-boot.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>
(cherry picked from commit 66daf04663cc8aa8c782df2036494383da62f522)