base: lmp-image-common: enable support for pam_mkhomedir
pam_mkhomedir can be used to automatically generate the user home
directory while creating the session. This module can be used together
with the factory reset functionality, which removes all available user
home directories.

The common-session change is done at image level as it also requires
the package to be installed, otherwise it breaks the login session.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>