base: mfgtool-resize-helper needs to be more forceful
The -p option on e2fsck is not enough to make sure the partition is clean for resizing
so change it to a -y so it will fix.

Then resize may still want e2fsck but since we just did it, we can force resize2fs to
do the resize anyway.

Signed-off-by: Tim Anderson <tim.anderson@foundries.io>
(cherry picked from commit daa50ca8e287f128c030d079a17c2faff6ada6bb)