base: initramfs-module-install-efi: give option to recreate ptable
Give the user an option to allow the user to decide if the installer
should recreate or not the partition table on the target storage device.

This is useful for scenarios in which the user can't simply just erase
and recreate the partition table from scratch as it might contain a
firmware / vendor specific partition required by the boot firmware.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>