base: u-boot-tools: drop CVE-2021-27138 fixes with @ fit separator
FIT_NODE_SEPARATOR allows the user to still use @ as fit separator, even
when we know that it is vulnerable to CVE-2021-27138, so make sure we
drop the required fixes when building u-boot-tools.

This is to allow customers to control when to transition to the correct
separator format '-' (default on new factories).

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>
(cherry picked from commit 80b773071cdc01b322bf4af142e32525037f467a)