base: lmp: enable readonly ostree sysroot
ostree supports mounting sysroot with readonly only v2020.1, and the
missing integration logic landed in meta-updater recently, so make sure
we also set as default in LmP.

This should help protecting against accidental damage when manipulating
sysroot. Ostree has the required logic to remount as rw before applying
an update, so it should work the same way as before from the user
perspective.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>