bsp: linux-tegra: add config fragment for lmp
Linux-tegra is not yet integrated with our base config fragments as the
kernel version used is quite old (before we created lmp-kernel-cache),
so just extend the recipe to include a config fragment used to extend
the base configuration with configs relevant to lmp.

Fragment initially with CONFIG_CRYPTO_LZ4 only, which is needed for zram
support.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>