base: init-install-efi: don't abort on efibootmgr failures
While we're checking for /sys/firmware/efi/efivars before running
efibootmgr, it might not necessarily work depending on the target
machine used (e.g. qemu), so make sure to continue the install process
even on failures.

This is not critical as the efi payload is installed using the generic
arm64 naming.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>