base: u-boot-fio-common: change config to be either lmp or lmp-ebbr
Currently the lmp-ebbr config fragment is being used together with lmp,
which doesn't work right on targets that are reducing the default set of
configs via the standard lmp fragment (for security).

As EBBR requires a large set of features to be enabled by default,
change the current logic to use either lmp-ebbr or lmp, not both.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>