base: docker-moby: patch to fsync layer metadata
Fsync `link` and `lower` layer files just after write to make sure their
content is stored on a persistent storage. It helps to overcome
the issue that occurs when image layers download/extraction is
interrupted by a power cut.
Docker daemon can recover if the interrupted layer is "broken",
but it cannot recover if there are two or more lower to the interrupted
layers are "broken". The given fix makes sure that the lower layers that
are fully downloaded and extracted are consistent even if the current
top layer is broken.

Signed-off-by: Mike Sul <mike.sul@foundries.io>
(cherry picked from commit 86d02315bd73dbd99d7fe42adb05ac134dde1008)