base: networkmanager: add patch to avoid managing docker bridges
Bring back the changes done by
https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/merge_requests/15
by forcing docker bridges to be unmanaged by network-manager.

This is because while they are normal externally managed bridges (and
the cause of the revert of the original change), they can be disabled
and won't be enabled automatically while turning network off and on via
nmcli.

This should be raised with upstream as a bug for finding out a proper
fix.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>