base: systemd: define default packageconfig via bbappend
To allow the user to enable support for networkd, we can't really use
the _removal syntax because that always gets parsed last, so instead
define a default set of packageconfig entries used and required by the
default LmP configuration via bbappends.

This is heavily based on the oe-core systemd recipe, and will have to be
maintained and updated as the main recipe adds/removes new options.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>