lmp: default image format should really be a default
Forcing the IMAGE_FSTYPES is bad because it overrides the BSP actual
default values, which is more likely to be appropriate for the device than
the one provides by lmp.

Doing so forces the BSP adaptation to IMAGE_FSTYPES:append to restore a
sensible default. This brings another problem with initramfs which tries to
reset the IMAGE_FSTYPES for a good reason. Because of the append, another
image type is added to the initramfs, when all we added is probably a cpio.

A good example of that in meta-lmp-bsp is tegra which needs "tegraflash"
for the actual image and because it is added through an append, it needs to
removed from the initramfs. Such approach, while kinda working, really
don't scale well. It would be better to let the BSP (and its adaptation to
LmP) deal with the default fstype.

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>