bsp: imx-mkimage: drop SPL-only signing patch
This patch was a work around for our signing script which was catenating the CSF binary
to the SPL/imx-boot.  This has been corrected now to insert at the correct location.

SoCs such as i.MX8M Plus and Nano require the additional CSF padding space even when not signed.

Signed-off-by: Michael Scott <mike@foundries.io>