bsp: imx-mkimage: drop SPL-only signing patch
This patch was a work around for our signing script which was catenating the CSF binary
to the SPL/imx-boot.  This has been corrected now to insert at the correct location.

SoCs such as i.MX8M Plus and Nano require the additional CSF padding space even when not signed.

Relevant change in lmp-tools (signing script) that supports this change:
https://github.com/foundriesio/lmp-tools/commit/27f6f61ad7e901454411d4749456e011c1b2aeef

Signed-off-by: Michael Scott <mike@foundries.io>