base: linux-lmp: skip debug split with modsign
Due the kernel module signatures we have to skip module stripping and
also debug split, as that will manipulate the module and invalidate the
signature.

This is now needed because of oe-core e09a8fa931, which added support
for kernel stripping in package and removed the special handling for ko
files.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>