Revert "bsp: lmp-machine-custom: am64xx: set different tmpdir for k3r5"
This reverts commit 05ad4773bba0975260e934fa41504b53c7fa6166.

Using an override for TMPDIR doesn't work on kirkstone,
maybe because it is applied at parse time.
Reverting [1] solves this but is not the correct solution.

| ERROR: mc:k3r5:binutils-cross-arm-2.38-r0 do_patch: Manifest /tmp/ti/build-lmp/tmp-k3r5/sstate-control/manifest-x86_64_ubuntu-20.04-quilt-native.populate_sysroot not found in x86_64 x86_64_ubuntu-20.04 (variant '')?
| ERROR: Logfile of failure stored in: /tmp/ti/build-lmp/tmp-k3r5/work/x86_64-linux/binutils-cross-arm/2.38-r0/temp/log.do_patch.71094
| ERROR: Task (mc:k3r5:/tmp/ti/build-lmp/conf/../../layers/openembedded-core/meta/recipes-devtools/binutils/binutils-cross_2.38.bb:do_patch) failed with exit code '1'

Reverting and defining the TMPDIR in the bitbake multi-config configuration
works as expected. The proposed solution [2] is not accepted upstream in
meta-ti and they said multiconfig can works perfectly without any issue in
the same TMPDIR.

Using the same TMPDIR for multicongig works but it has some side effects
mainly in OE-core archiver and rm_work bbclasses that have been identified
and fixed.

[1] https://git.openembedded.org/openembedded-core/commit/?id=64b89f3c8fc31842256c482a3039d90d3f12c1cc
[2] https://lists.yoctoproject.org/g/meta-ti/message/14767

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>