base: ostree: add patch to support boot under a directory
Allow manipulating and updating /boot/loader entries under a normal
directory, as well as using symbolic links (default).

For directories this uses `renameat2` to do atomic swap of the loader
directory in the boot partition. It fallsback to non-atomic rename.
This stays atomic on filesystems supporting links but also provide
a non-atomic behavior when filesystem does not provide any atomic
alternative.

/boot/loader as a normal directory is needed for systemd-boot support,
and it needs to be stored under the EFI ESP vfat partition.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>