base/bsp: openssh: drop rng-tools from default PACKAGECONFIG
Since kernel 5.6 reading /dev/random does not block anymore, and CRNG
(initialized by the kernel during boot) is considered to be good enough
for userspace, so no need for having rng-tools seeding entropy.

This is not done with a remove override as we still want to allow BSP
layers (or via meta-lmp-bsp) to be able to add rng-tools when needed
(e.g. kernel older than 5.6, as done with tegra and apalis-imx8).

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>