aktualizr: start service after time is synced
On boot, the start of aktualizr-lite tends to be deplayed by 5min.
The time might not be synced yet causing the certificate validation to
fail. Such issue is probably only seen on system witout an RTC, which may
wake up with extravagant time.

Making aktualizr-lite.service wait for time-sync.target solves the problem.

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>