base: fip-utils: simplify the fip signing proces
Use same key as used by TF-A when generating ROT_HASH (TF_A_SIGN_KEY),
and simplify the process by removing the search path logic and assuming
that the user is giving a key with the absolute path instead (as done
for the other u-boot/kernel keys currently used in lmp).

Suffix also removed as a way to simplify the image/wic/flashing process.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>