bsp: u-boot-fio: simplify deploy logic for stm32
Currently only relevant for cases where u-boot signing is not required,
as by default u-boot won't deploy the dtb and nodtb binaries used, and
which are required for fip create.

This is similar to the logic done for imx8 devices, so we can probably
have something done in a more generic way later.

Signed-off-by: Ricardo Salveti <ricardo@foundries.io>