lmp-staging: serialize do_compile of heavy tasks
This is a work around to not compile rust and clang at same
time as it is very huge tasks.

It uses the bitbake lockfiles to run the refered tasks one at a time.
The bitbake UI can be misleading since it will inform you that the tasks
are working at the same time, which is not true because only one is
working and all the others are blocked.
I have a pending bitbake patch to improve this situation.

Signed-off-by: Jose Quaresma <jose.quaresma@foundries.io>